+++ tests/project_items/importer/test_Importer.py
@@ -78,7 +78,7 @@ class TestImporter(unittest.TestCase):
         )
         source_item.item_type = MagicMock(return_value="Data Store")
         self.importer.notify_destination(source_item)
-        self.toolbox.msg.emit.assert_called_with("Link established.")
+        self.toolbox.msg.emit.assert_called_with("Link established")
         source_item.item_type = MagicMock(return_value="Exporter")
         self.importer.notify_destination(source_item)
         self.toolbox.msg_warning.emit.assert_called_with(